pyproject.toml
[project]
name = "generate_cards"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Martin Faucheux", email = "martin.faucheux@gmail.com" }]
dependencies = [
    "google-genai>=1.38.0",
    "python-dotenv>=1.0.0",
    "notion-client>=2.2.1",
    "requests>=2.31.0",
    "pillow>=11.3.0",
]
requires-python = "==3.13.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false


[tool.pdm.scripts]
generate = "python run.py new"
modify = "python run.py modify"
